+++ packages/zowe-explorer/src/trees/dataset/DatasetTree.ts
@@ -15,6 +15,7 @@ import dayjs from "dayjs";
 import {
     Gui,
     handleError,
+    handleError,
     Validation,
     imperative,
     IZoweDatasetTreeNode,