+++ packages/zowe-explorer/src/trees/job/JobTree.ts
@@ -15,6 +15,7 @@ import { IJob } from "@zowe/zos-jobs-for-zowe-sdk";
 import {
     Gui,
     handleError,
+    handleError,
     Validation,
     imperative,
     IZoweJobTreeNode,