+++ packages/zowe-explorer/src/trees/shared/SharedInit.ts
@@ -13,6 +13,7 @@ import * as vscode from "vscode";
 import {
     FileManagement,
     handleError,
+    handleError,
     Gui,
     CorrelatedError,
     IZoweTree,