+++ packages/zowe-explorer/src/trees/uss/USSTree.ts
@@ -13,6 +13,8 @@ import * as vscode from "vscode";
 import * as path from "path";
 import {
     errorMessage,
+    handleError,
+    handleError,
     FsAbstractUtils,
     Gui,
     imperative,