+++ packages/zowe-explorer/src/trees/uss/UssFSProvider.ts
@@ -14,6 +14,7 @@ import * as vscode from "vscode";
 import {
     BaseProvider,
     handleError,
+    handleError,
     BufferBuilder,
     FsAbstractUtils,
     imperative,